Contador de reservas me dice cuantas reservas tiene el cliente en todo momento Lo agrege a la barra

+++ servidor-web/src/main/java/uy/edu/cure/servidor/web/CarritoBean.java
@@ -40,7 +40,7 @@ public class CarritoBean implements Serializable{
         item.setFechaInicio(fechaInicio);
         item.setFechaFin(fechaFin);
         item.setProducto(servicioBean.getServicio());
-        itemCantidad = (Double) item.getCantidad();
+        itemCantidad = item.getCantidad();
         itemPrecio =  item.getProducto().getPrecio();
         item.setSubTotal(itemCantidad * itemPrecio);
         CarritoController carritofinal = new CarritoController();
